requirements.txt
Django>=4.2
psycopg2-binary
requests
python-dotenv
Pillow
django-crispy-forms
crispy-bootstrap5
django-apscheduler
